+++ frontend/src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from "@angular/core";
 import { RouterModule, Routes } from "@angular/router";
 
-import { PostDetailComponent } from "./post-detail/post-detail.component";
-import { PostsComponent } from "./posts/posts.component";
-import { LoginComponent } from "./login/login.component";
-import { RegisterComponent } from "./register/register.component";
-import { CreatePostComponent } from "./create-post/create-post.component";
+import { PostDetailComponent } from "./components/post-detail/post-detail.component";
+import { PostsComponent } from "./components/posts/posts.component";
+import { LoginComponent } from "./components/login/login.component";
+import { RegisterComponent } from "./components/register/register.component";
+import { CreatePostComponent } from "./components/create-post/create-post.component";
 
 const routes: Routes = [
   { path: "", redirectTo: "/posts", pathMatch: "full" },
