+++ frontend/src/app/components/posts/posts.component.html
@@ -1,8 +1,6 @@
 <h2>Rooms for rent</h2>
 <app-filter></app-filter>
-<label>Search:
-  <input type="button" value="search">
-</label>
+
 <br>
 <label>Area:
   <input type="checkbox" value="sydney"> Sydney &nbsp;
