+++ frontend/src/app/components/posts/posts.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from "@angular/core";
-import { Post } from "../post";
-import { PostService } from "../post.service";
+import { Post } from "../../post";
+import { PostService } from "../../post.service";
 
 @Component({
   selector: "app-posts",
